add VLANs and NetworkInterfaces

+++ src/main/java/pl/lodz/p/repository/NetworkInterfaceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pl.lodz.p.domain.entities.NetworkInterface;
 
-public interface NetworkInterfaceRepository extends JpaRepository<NetworkInterface, Integer> {
+public interface NetworkInterfaceRepository extends JpaRepository<NetworkInterface, Long> {
 
 }
