+++ src/main/java/pl/lodz/p/repository/NetworkServiceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pl.lodz.p.domain.entities.NetworkService;
 
-public interface NetworkServiceRepository extends JpaRepository<NetworkService, Integer> {
+public interface NetworkServiceRepository extends JpaRepository<NetworkService, Long> {
 
 }
