+++ src/main/java/pl/lodz/p/repository/VLANRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pl.lodz.p.domain.entities.VLAN;
 
-public interface VLANRepository extends JpaRepository<VLAN, Integer> {
+public interface VLANRepository extends JpaRepository<VLAN, Long> {
 
 }
